requirements.txt
pytorch
pyg
numpy
